♻️ refactor(aiAgent): extract top-level helpers and shared types from index.ts
Split the 6873-line AiAgentService index.ts by moving pure top-level
helpers into dedicated modules:

- helpers/agentFactory.ts: createGraphAwareAgentFactory, STOPPED_TOOL_CONTENT
- helpers/groupContext.ts: buildGroupAgentContext, buildBotConversationGroupContext, formatErrorForMetadata
- helpers/mediaAvailability.ts: media availability probes, isMultimodalUnderstandingConfigured
- helpers/heteroErrors.ts: hetero dispatch error humanizing and type mapping
- types.ts: InternalExecAgentParams, ResolvedWorkspaceInit

Pure code motion with zero behavior change. Public exports unchanged.

LOBE-13709
